requirements.txt
Jinja2==3.0.1
Pygments==2.9.0
Quandl==3.6.1
azure-identity==1.6.0
azure-storage-blob==12.8.1
pycapnp==1.1.0
coinbasepro==0.2.1
cryptofeed==1.9.1
dask[complete]==2021.6.1
diskcache==5.2.1
fire==0.4.0
hvac==0.10.14
luigi==3.0.3
money==1.3.0
msal==1.12.0
numpy==1.20.3
pandas==1.2.4
pandas_market_calendars==2.0
phemex==0.3.0
prometheus-client==0.11.0
psycopg2-binary==2.9.1
py-consul==1.2.3
pyarrow==4.0.1
pyfolio==0.9.2
pytau==0.14.0
pytest-mock==3.6.1
python-binance==1.0.12
pytz==2021.1
pyyaml==5.4.1
pyzmq==22.1.0
requests==2.25.1
scipy==1.6.3
sqlalchemy==1.4.18
tables==3.6.1
toml==0.10.2
websockets==9.1
yfinance==0.1.59